corrected returns

+++ 0x15-file_io/0-read_textfile.c
@@ -14,7 +14,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int fd;
+	ssize_t fd;
 	char *buff;
 	ssize_t fd_write, n_bytes;
